Support manifests with an array of applications
The Cloud Foundry manifest have two possible formats: one with only one
app and, optionally, a list of applications under the key 'applications'

The optional `environment_variables` assumes that the manifest is using
only the first format, so it is not able to change the variables in the
second format.

With this commit we check the format that the manifest is define, and we
modify the value of the `env` hash for the first app in the list.